requirements.txt
asgiref==3.8.1
boto3==1.35.41
botocore==1.35.41
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-phonenumber-field==8.0.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
jmespath==1.0.1
newsapi-python==0.2.7
numpy==2.1.2
packaging==24.1
phonenumbers==8.13.47
pillow==10.4.0
psutil==6.0.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
